Move gmtime/localtime tests to test_time.cpp
