--- tests/core/test_gmtime_localtime.out
@@ -1,8 +1,0 @@
-time: 0, gmtime: 1970-01-01 00:00:00, localtime: 1969-12-31 19:00:00
-time: 2147483647, gmtime: 2038-01-19 03:14:07, localtime: 2038-01-18 22:14:07
-time: 2147483648, gmtime: 2038-01-19 03:14:08, localtime: 2038-01-18 22:14:08
-time: -2147483648, gmtime: 1901-12-13 20:45:52, localtime: 1901-12-13 15:45:52
-time: -2147483649, gmtime: 1901-12-13 20:45:51, localtime: 1901-12-13 15:45:51
-time: 253402300799, gmtime: 9999-12-31 23:59:59, localtime: 9999-12-31 18:59:59
-time: -62135578800, gmtime: 1-01-01 05:00:00, localtime: 1-01-01 00:00:00
-success
