+++ tests/core/test_time.out
@@ -51,3 +51,11 @@ timespec_get test 3: 1
 timespec_get test 4: 1
 timespec_get test 5: 1
 timespec_get test 6: 1
+time: 0, gmtime: 1970-01-01 00:00:00
+time: 2147483647, gmtime: 2038-01-19 03:14:07
+time: 2147483648, gmtime: 2038-01-19 03:14:08
+time: -2147483648, gmtime: 1901-12-13 20:45:52
+time: -2147483649, gmtime: 1901-12-13 20:45:51
+time: 253402300799, gmtime: 9999-12-31 23:59:59
+time: -62135596800, gmtime: 1-01-01 00:00:00
+success
